porting over cmnist from nosinn
