+++ ethicml/data/__init__.py
@@ -1,16 +1,1 @@
 """This module contains items related to data, such as raw csv's and data objects."""
-
-from .adult import *
-from .celeba import *
-from .compas import Compas
-from .configurable_dataset import *
-from .credit import *
-from .dataset import *
-from .genfaces import *
-from .german import *
-from .image_dataset import *
-from .load import *
-from .non_binary_toy import *
-from .sqf import *
-from .toy import *
-from .util import *
