+++ ethicml/data/csvs/make_health_from_raw.py
@@ -3,7 +3,7 @@
 
 import pandas as pd
 
-df = pd.read_csv("./raw/health.csv")
+df = pd.read_csv("raw/health.csv")
 
 # Drop columns which contains NaNs
 print(f"Dropping columns: {df.columns[df.isna().any()].tolist()}")
