--- ethicml/data/health.py
@@ -1,309 +1,0 @@
-"""Class to describe features of the Heritage Health dataset."""
-from ethicml.common import implements
-from ethicml.data import Dataset
-
-__all__ = ["Health"]
-
-
-class Health(Dataset):
-    """Heritage Health Dataset."""
-
-    @property
-    def name(self) -> str:
-        """Getter for dataset name."""
-        return "Health"
-
-    @property
-    def filename(self) -> str:
-        """Getter for filename."""
-        return "health.csv.zip"
-
-    @implements(Dataset)
-    def __len__(self) -> int:
-        return 171067
-
-    def __init__(self, split: str = "Sex", discrete_only: bool = False):
-        """Init Heritage Health dataset."""
-        super().__init__()
-        self.split = split
-        self.discrete_only = discrete_only
-        self.features = [
-            "MemberID_t",
-            "YEAR_t",
-            "ClaimsTruncated",
-            "trainset",
-            "age_05",
-            "age_15",
-            "age_25",
-            "age_35",
-            "age_45",
-            "age_55",
-            "age_65",
-            "age_75",
-            "age_85",
-            "age_MISS",
-            "sexMALE",
-            "sexFEMALE",
-            "sexMISS",
-            "no_Claims",
-            "no_Providers",
-            "no_Vendors",
-            "no_PCPs",
-            "no_PlaceSvcs",
-            "no_Specialities",
-            "no_PrimaryConditionGroups",
-            "no_ProcedureGroups",
-            "PayDelay_max",
-            "PayDelay_min",
-            "PayDelay_ave",
-            "PayDelay_stdev",
-            "LOS_max",
-            "LOS_min",
-            "LOS_ave",
-            "LOS_stdev",
-            "LOS_TOT_UNKNOWN",
-            "LOS_TOT_SUPRESSED",
-            "LOS_TOT_KNOWN",
-            "dsfs_max",
-            "dsfs_min",
-            "dsfs_range",
-            "dsfs_ave",
-            "dsfs_stdev",
-            "CharlsonIndexI_max",
-            "CharlsonIndexI_min",
-            "CharlsonIndexI_ave",
-            "CharlsonIndexI_range",
-            "CharlsonIndexI_stdev",
-            "pcg1",
-            "pcg2",
-            "pcg3",
-            "pcg4",
-            "pcg5",
-            "pcg6",
-            "pcg7",
-            "pcg8",
-            "pcg9",
-            "pcg10",
-            "pcg11",
-            "pcg12",
-            "pcg13",
-            "pcg14",
-            "pcg15",
-            "pcg16",
-            "pcg17",
-            "pcg18",
-            "pcg19",
-            "pcg20",
-            "pcg21",
-            "pcg22",
-            "pcg23",
-            "pcg24",
-            "pcg25",
-            "pcg26",
-            "pcg27",
-            "pcg28",
-            "pcg29",
-            "pcg30",
-            "pcg31",
-            "pcg32",
-            "pcg33",
-            "pcg34",
-            "pcg35",
-            "pcg36",
-            "pcg37",
-            "pcg38",
-            "pcg39",
-            "pcg40",
-            "pcg41",
-            "pcg42",
-            "pcg43",
-            "pcg44",
-            "pcg45",
-            "pcg46",
-            "sp1",
-            "sp2",
-            "sp3",
-            "sp4",
-            "sp5",
-            "sp6",
-            "sp7",
-            "sp8",
-            "sp9",
-            "sp10",
-            "sp11",
-            "sp12",
-            "sp13",
-            "pg1",
-            "pg2",
-            "pg3",
-            "pg4",
-            "pg5",
-            "pg6",
-            "pg7",
-            "pg8",
-            "pg9",
-            "pg10",
-            "pg11",
-            "pg12",
-            "pg13",
-            "pg14",
-            "pg15",
-            "pg16",
-            "pg17",
-            "pg18",
-            "ps1",
-            "ps2",
-            "ps3",
-            "ps4",
-            "ps5",
-            "ps6",
-            "ps7",
-            "ps8",
-            "ps9",
-            "drugCount_max",
-            "drugCount_min",
-            "drugCount_ave",
-            "drugcount_months",
-            "labCount_max",
-            "labCount_min",
-            "labCount_ave",
-            "labcount_months",
-            "labNull",
-            "drugNull",
-        ]
-
-        self.features_to_remove = ["MemberID_t", "YEAR_t", "trainset", "sexMISS", "age_MISS"]
-
-        self.continuous_features = [
-            "no_Claims",
-            "no_Providers",
-            "no_Vendors",
-            "no_PCPs",
-            "no_PlaceSvcs",
-            "no_Specialities",
-            "no_PrimaryConditionGroups",
-            "no_ProcedureGroups",
-            "PayDelay_max",
-            "PayDelay_min",
-            "PayDelay_ave",
-            "PayDelay_stdev",
-            "LOS_max",
-            "LOS_min",
-            "LOS_ave",
-            "LOS_stdev",
-            "LOS_TOT_UNKNOWN",
-            "LOS_TOT_SUPRESSED",
-            "LOS_TOT_KNOWN",
-            "dsfs_max",
-            "dsfs_min",
-            "dsfs_range",
-            "dsfs_ave",
-            "dsfs_stdev",
-            "CharlsonIndexI_max",
-            "CharlsonIndexI_min",
-            "CharlsonIndexI_ave",
-            "CharlsonIndexI_range",
-            "CharlsonIndexI_stdev",
-            "pcg1",
-            "pcg2",
-            "pcg3",
-            "pcg4",
-            "pcg5",
-            "pcg6",
-            "pcg7",
-            "pcg8",
-            "pcg9",
-            "pcg10",
-            "pcg11",
-            "pcg12",
-            "pcg13",
-            "pcg14",
-            "pcg15",
-            "pcg16",
-            "pcg17",
-            "pcg18",
-            "pcg19",
-            "pcg20",
-            "pcg21",
-            "pcg22",
-            "pcg23",
-            "pcg24",
-            "pcg25",
-            "pcg26",
-            "pcg27",
-            "pcg28",
-            "pcg29",
-            "pcg30",
-            "pcg31",
-            "pcg32",
-            "pcg33",
-            "pcg34",
-            "pcg35",
-            "pcg36",
-            "pcg37",
-            "pcg38",
-            "pcg39",
-            "pcg40",
-            "pcg41",
-            "pcg42",
-            "pcg43",
-            "pcg44",
-            "pcg45",
-            "pcg46",
-            "sp1",
-            "sp2",
-            "sp3",
-            "sp4",
-            "sp5",
-            "sp6",
-            "sp7",
-            "sp8",
-            "sp9",
-            "sp10",
-            "sp11",
-            "sp12",
-            "sp13",
-            "pg1",
-            "pg2",
-            "pg3",
-            "pg4",
-            "pg5",
-            "pg6",
-            "pg7",
-            "pg8",
-            "pg9",
-            "pg10",
-            "pg11",
-            "pg12",
-            "pg13",
-            "pg14",
-            "pg15",
-            "pg16",
-            "pg17",
-            "pg18",
-            "ps1",
-            "ps2",
-            "ps3",
-            "ps4",
-            "ps5",
-            "ps6",
-            "ps7",
-            "ps8",
-            "ps9",
-            "drugCount_max",
-            "drugCount_min",
-            "drugCount_ave",
-            "drugcount_months",
-            "labCount_max",
-            "labCount_min",
-            "labCount_ave",
-            "labcount_months",
-        ]
-
-        if split == "Sex":
-            self.sens_attrs = ["sexMALE"]
-            self.s_prefix = ["sex"]
-            self.class_labels = ["Charlson>0"]
-            self.class_label_prefix = ["Charlson"]
-        else:
-            raise NotImplementedError
